# Fixes #16010 - Include referenced objects in hosts