@akofink 2016-09-22
Waiting on Foreman to implement an extension http://projects.theforeman.org/issues/16646