@akofink 2016-09-26
Tests won't pass without https://github.com/theforeman/foreman/pull/3887